requirements.txt
gradio==4.44.0
tensorflow-cpu==2.15.0
numpy==1.26.4